More documentation fixes.
Changed seline()s signature. The previous one was too hard to even define
strictly because the length of a small diagonal element should be multiplied
by \sqrt{2} but at the same time it is not clear how you should define where the
line ends and then you have to round it and it gets ugly.

This is dumber and better.